+++ src/timer.rs
@@ -29,6 +29,12 @@ impl SysTimer {
     }
 }
 
+impl Default for SysTimer {
+    fn default() -> SysTimer {
+        SysTimer::new()
+    }
+}
+
 impl CountDown for SysTimer {
     type Error = Infallible;
     type Time = Duration;